+++ pages/_campus/users/index.vue
@@ -56,6 +56,7 @@
           :can-create-new="$auth.isAdmin(campus.id)"
           export-csv
           :mask="mask"
+          has-mask
           :pagination="usersPagination"
           :api-call="$api.campusUsers(campus.id, mask).getUsers"
           @importCSV="usersUploadCSV"
@@ -120,7 +121,7 @@ export default {
     const driversReq = $api.drivers(params.campus, Object.keys(columns).join(','))
       .getDrivers({ offset, limit });
     const usersReq = $api.campusUsers(params.campus, Object.keys(columns).join(','))
-      .getUsers(offset, limit);
+      .getUsers({ offset, limit });
     const driversRes = await driversReq;
     const usersRes = await usersReq;
     return {
@@ -132,7 +133,7 @@ export default {
   },
   data() {
     return {
-      mask: 'id,email,firstname,lastname,phone',
+      mask: 'email,firstname,lastname,phone(original,canonical)',
     };
   },
   computed: {